+++ polidoro_argument/__init__.py
@@ -1,10 +1,3 @@
-"""
-A Module to make easier to create script with command line arguments
-See README.md for more information
-"""
 from polidoro_argument.argument import Argument
+from polidoro_argument.argument_action import ArgumentAction
 from polidoro_argument.argument_parser import ArgumentParser
-from polidoro_argument.command import Command
-
-NAME = 'py_argument'
-VERSION = '3.0.0'
