+++ polidoro_argument/argument.py
@@ -1,42 +1,62 @@
 """
-A Module to make easier to create script with command line arguments.
-See README.md for more information
+Decorator to add an function/method as argument in parser
 """
 import inspect
 
-from polidoro_argument.custom_action import CustomAction
+from polidoro_argument.argument_action import ArgumentAction
+from polidoro_argument.argument_parser import ArgumentParser
 
 
-class Argument:  # pylint: disable=too-few-public-methods
-    """
-    Decorator to create a command line argument
-    """
-    arguments = []
-
-    def __init__(self, method=None, **kwargs):
-        self.method = method
-        # if method is None the decorator has parameters
-        if method is None:
-            self.kwargs = kwargs
+class Argument(object):
+    def __new__(cls, method=None, **kwargs):
+        if callable(method):
+            # When the decorator has no arguments
+            # noinspection PyTypeChecker
+            return Argument._add_argument(method, **kwargs)
         else:
-            self.args = tuple(['--' + method.__name__])
+            # When the decorator has arguments
+            def wrapper(_method):
+                return Argument._add_argument(_method, **kwargs)
 
-            parameters = [p for p in inspect.signature(method).parameters if not p.startswith('_')]
-            # nargs = number of arguments in method
-            self.kwargs = {
-                'action': CustomAction,
-                'method': method,
-                'nargs': len(parameters)
-            }
-            if parameters:
-                self.kwargs['metavar'] = ' '.join(parameters)
+            return wrapper
 
-            self.arguments.append(self)
+    @staticmethod
+    def _add_argument(method, **kwargs):
+        if not hasattr(method, '__name__'):
+            raise RuntimeError('The "method" must have the attribute "__name__"!')
+        """ Add the method as argument in parser, to be callas as --METHOD_NAME """
+        parser = ArgumentParser.get_parser()
 
-    def __call__(self, *args, **kwargs):
-        # if method is None the decorator has parameters and the first argument is the method
-        if self.method is None:
-            self.method = list(args)[0]
-            self.__init__(**vars(self))
-            return self
-        return self.method(*args, **kwargs)
+        parameters = {name: info for name, info in inspect.signature(method).parameters.items() if
+                      not name.startswith('_')}
+        required_parameters = []
+        optional_parameters = []
+        for name, info in parameters.items():
+            # noinspection PyUnresolvedReferences,PyProtectedMember
+            if info.default == inspect._empty:
+                required_parameters.append(name)
+            else:
+                optional_parameters.append(name)
+        # nargs = number of arguments in method
+        nargs_min = len(required_parameters)
+        nargs_max = len(required_parameters + optional_parameters)
+        metavar = tuple(parameters)
+        if nargs_min == nargs_max:
+            nargs = nargs_min
+        else:
+            nargs = '*'
+            # metavar = ' '.join(required_parameters + ['[%s]' % opt_param for opt_param in optional_parameters])
+        kwargs.update({
+            'action': ArgumentAction,
+            'method': method,
+            'nargs_min': nargs_min,
+            'nargs_max': nargs_max,
+            'nargs': nargs,
+            'metavar': metavar,
+            'required_parameters': required_parameters,
+            'optional_parameters': optional_parameters,
+        })
+
+        parser.add_argument('--' + method.__name__, **kwargs)
+
+        return method
