+++ polidoro_argument_old/__init__.py
@@ -1,0 +1,10 @@
+"""
+A Module to make easier to create script with command line arguments
+See README.md for more information
+"""
+from polidoro_argument_old.argument import Argument
+from polidoro_argument_old.argument_parser import ArgumentParser
+from polidoro_argument_old.command import Command
+
+NAME = 'py_argument'
+VERSION = '3.0.0'
