+++ setup.py
@@ -3,18 +3,18 @@ Setup to create the package
 """
 import setuptools
 
-import polidoro_argument
+import polidoro_argument_old
 
 with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name='polidoro-py-argument',
-    version=polidoro_argument.VERSION,
+    name='polidoro-argument',
+    version=polidoro_argument_old.VERSION,
     description='Package to create command line arguments for Python.',
     long_description=long_description,
     long_description_content_type="text/markdown",
-    url='https://github.com/heitorpolidoro/py-argument',
+    url='https://github.com/heitorpolidoro/polidoro-argument',
     author='Heitor Polidoro',
     # author_email='heitor.polidoro',
     license='unlicense',
