+++ tests/argument_with_parameters_test.py
@@ -4,7 +4,7 @@ Test file
 
 import pytest
 
-from pyargument import Argument, ArgumentParser
+from polidoro_argument import Argument, ArgumentParser
 
 ArgumentParser._parser = None
